+++ example/lib/models/class.dart
@@ -50,4 +50,9 @@ class Class extends RelationshipModel with OneToOne, OneToMany, ManyToMany {
         'createdAt': createdAt?.toIso8601String(),
         'updatedAt': updatedAt?.toIso8601String()
       };
+
+  @override
+  setPrimaryValue(value) {
+    id = value;
+  }
 }