+++ example/lib/models/user.dart
@@ -77,4 +77,8 @@ class User extends RelationshipModel with OneToOne, OneToMany, ManyToMany {
         'createdAt': createdAt?.toIso8601String(),
         'updatedAt': updatedAt?.toIso8601String()
       };
+  @override
+  setPrimaryValue(value) {
+    id = value;
+  }
 }