+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+## 0.0.3+1
+
+- `save` method in model now act like `updateOrCreate` method.
+
+## 0.0.3
+
+- Added `RelationshipModel` for relationships.
+- Added `ManyToMany` mixin.
+- Updated doc.
+
 ## 0.0.2
 
 - Added `whereIn` method.